requirements.txt
appdirs==1.4.4
asgiref==3.3.0
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
distlib==0.3.1
distro==1.5.0
Django==3.1.2
djangorestframework==3.12.1
filelock==3.0.12
html5lib==1.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==1.0.0
packaging==20.4
pep517==0.9.1
pluggy==0.13.1
progress==1.5
py==1.9.0
pyparsing==2.4.7
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
rabbit==1.2.0
requests==2.24.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
tox==3.20.1
urllib3==1.25.11
virtualenv==20.1.0
webencodings==0.5.1
psycopg2==2.8.6